added new methods for characters

+++ OOPTeam2/RPG_Game/Models/Wands/Wand.cs
@@ -1,7 +1,7 @@
 ﻿namespace OOPTeam2.RPG_Game.Models.Wands {
     public abstract class Wand {
-        public double powerCoefficient { set; get; }
-        public int agility { set; get; }
+        public int powerCoefficient { set; get; }
+        public double agility { set; get; }
         public string description { set; get; }
         public bool isAvailable { set; get; }
